refactor/#55: experience->exp 이름 변경
