+++ app/(main)/exp/components/DropdownMenu.tsx
@@ -5,7 +5,7 @@ import TrashFullIcon from '@/public/icons/Trash_Full.svg';
 import EditPencilIcon from '@/public/icons/Edit_Pencil_02.svg';
 import { useRouter } from 'next/navigation';
 import Popup from '@/app/components/Popup';
-import { deleteExperience } from '@/apis/exp';
+import { deleteExp } from '@/apis/exp';
 
 interface DropdownMenuProps {
   id: number;
@@ -14,7 +14,7 @@ interface DropdownMenuProps {
 
 export default function DropdownMenu({ id, onClose }: DropdownMenuProps) {
   const menuRef = useRef<HTMLDivElement>(null);
-  const [form, setForm] = useState({ isPopupOpen: false });
+  const [isPopupOpen, setIsPopupOpen] = useState(false);
   const router = useRouter();
 
   useEffect(() => {
@@ -38,7 +38,7 @@ export default function DropdownMenu({ id, onClose }: DropdownMenuProps) {
       <button
         onClick={() => {
           onClose();
-          router.push(`/addExperience?id=${id}`);
+          router.push(`/addExp?id=${id}`);
         }}
         className="w-full text-left px-4 py-2 hover:bg-gray-300 flex items-center gap-2"
       >
@@ -46,23 +46,23 @@ export default function DropdownMenu({ id, onClose }: DropdownMenuProps) {
         수정
       </button>
       <button
-        onClick={() => setForm(prev => ({ ...prev, isPopupOpen: true }))}
+        onClick={() => setIsPopupOpen(true)}
         className="w-full text-left px-4 py-2 hover:bg-gray-300 flex items-center gap-2"
       >
         <TrashFullIcon className="stroke-gray-1000" />
         삭제
       </button>
-      {form.isPopupOpen && (
+      {isPopupOpen && (
         <Popup
           title="경험 삭제"
           content={`경험을 삭제하시겠습니까?\n삭제하시면 다시 복구할 수 없습니다.`}
           confirmText="삭제"
           cancelText="취소"
           onConfirm={async () => {
-            await deleteExperience(id);
+            await deleteExp(id);
             router.refresh();
           }}
-          onCancel={() => setForm(prev => ({ ...prev, isPopupOpen: false }))}
+          onCancel={() => setIsPopupOpen(false)}
         />
       )}
     </div>
