+++ app/(main)/exp/components/ExpList.tsx
@@ -1,22 +1,22 @@
 'use client';
 
 import { useState } from 'react';
-import { Experience } from '@/apis/exp';
-import { ExperienceType } from '@/types/exp';
-import ExperienceCard from './ExperienceCard';
+import { Exp } from '@/apis/exp';
+import { ExpType } from '@/types/exp';
+import ExpCard from './ExpCard';
 import BtnExp from '@/app/components/BtnExp';
 import BtnFilter from '@/app/components/BtnFilter';
 
 interface ExpListProps {
-  data: Experience[];
+  data: Exp[];
 }
 
 export default function ExpList({ data }: ExpListProps) {
-  const [selectedType, setSelectedType] = useState<ExperienceType | null>(null);
+  const [selectedType, setSelectedType] = useState<ExpType | null>(null);
   const [localData] = useState(data);
 
   const filteredData = selectedType
-    ? localData?.filter(exp => exp.experienceType === selectedType)
+    ? localData?.filter(exp => exp.expType === selectedType)
     : localData;
 
   return (
@@ -24,7 +24,7 @@ export default function ExpList({ data }: ExpListProps) {
       <h1 className="text-[25px] font-bold mb-10 flex items-center justify-between">
         <span>내 경험</span>
         <div className="flex justify-end gap-4">
-          <BtnExp href="/addExperience " className="bg-primary-50 text-black">
+          <BtnExp href="/addExp " className="bg-primary-50 text-black">
             경험 추가
           </BtnExp>
           <BtnFilter onSelectType={setSelectedType} />
@@ -35,12 +35,12 @@ export default function ExpList({ data }: ExpListProps) {
         <div>경험이 존재하지 않습니다.</div>
       ) : (
         <div className="w-full flex flex-wrap space-x-[28px] space-y-[37px]">
-          {filteredData?.map(experience => (
-            <ExperienceCard
-              key={experience.id}
-              id={experience.id}
-              title={experience.title}
-              type={experience.experienceType}
+          {filteredData?.map(exp => (
+            <ExpCard
+              key={exp.id}
+              id={exp.id}
+              title={exp.title}
+              type={exp.expType}
               isTemp={false}
             />
           ))}
