+++ app/(main)/exp/page.tsx
@@ -1,8 +1,8 @@
-import { getMyExperience } from '@/apis/exp';
+import { getMyExp } from '@/apis/exp';
 import ExpList from './components/ExpList';
 
 export default async function ExpMainPage() {
-  const { httpStatus, data = [] } = await getMyExperience();
+  const { httpStatus, data = [] } = await getMyExp();
 
   if (httpStatus !== 200) {
     return <div>오류가 발생했습니다.</div>;
