+++ app/components/TopBar.tsx
@@ -6,7 +6,7 @@ import { usePathname } from 'next/navigation';
 
 export default function TopBar() {
   const pathname = usePathname();
-  const pageState = pathname === '/experience' ? 'main' : 'more';
+  const pageState = pathname === '/exp' ? 'main' : 'more';
 
   return (
     <div className="flex justify-between pl-[398px] pr-[130px] bg-gray-900 h-[100px] p-7">
